Split market setting and acquiring
99% of the MarketManager callers do not want to set the market.
In fact, it is done in only one place, the login screen. Better to only
have it available at that one place instead through another interface.
