+++ app/market/market-core/src/main/kotlin/com/hedvig/android/market/SetMarketUseCase.kt
@@ -1,0 +1,13 @@
+package com.hedvig.android.market
+
+interface SetMarketUseCase {
+  suspend fun setMarket(market: Market)
+}
+
+internal class SetMarketUseCaseImpl(
+  private val marketStorage: MarketStorage,
+) : SetMarketUseCase {
+  override suspend fun setMarket(market: Market) {
+    marketStorage.setMarket(market)
+  }
+}
