+++ app/market/market-core/src/main/kotlin/com/hedvig/android/market/di/MarketManagerModule.kt
@@ -6,8 +6,13 @@ import androidx.datastore.preferences.core.Preferences
 import com.hedvig.android.core.common.ApplicationScope
 import com.hedvig.android.market.MarketManager
 import com.hedvig.android.market.MarketManagerImpl
+import com.hedvig.android.market.MarketStorage
+import com.hedvig.android.market.SetMarketUseCase
+import com.hedvig.android.market.SetMarketUseCaseImpl
 import org.koin.dsl.module
 
 val marketManagerModule = module {
-  single<MarketManager> { MarketManagerImpl(get<DataStore<Preferences>>(), get<Context>(), get<ApplicationScope>()) }
+  single<MarketManager> { MarketManagerImpl(get<MarketStorage>(), get<ApplicationScope>()) }
+  single<MarketStorage> { MarketStorage(get<DataStore<Preferences>>(), get<Context>()) }
+  single<SetMarketUseCase> { SetMarketUseCaseImpl(get<MarketStorage>()) }
 }
